requirements.txt
django==4.0.3
# django-environ==0.8.1
django-bootstrap-v5==1.0.11
djangorestframework==3.13.1
django-cors-headers
django-phonenumber-field[phonenumbers]